+++ tests/bluetooth/common/testlib/include/testlib/att_read.h
@@ -2,6 +2,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+#ifndef ZEPHYR_TESTS_BLUETOOTH_COMMON_TESTLIB_INCLUDE_TESTLIB_ATT_READ_H_
+#define ZEPHYR_TESTS_BLUETOOTH_COMMON_TESTLIB_INCLUDE_TESTLIB_ATT_READ_H_
+
 #include <stdint.h>
 #include <zephyr/bluetooth/gatt.h>
 
@@ -20,9 +23,9 @@ int bt_testlib_att_read_by_handle_sync(struct net_buf_simple *result_data, uint1
 				       enum bt_att_chan_opt bearer, uint16_t handle,
 				       uint16_t offset);
 
-int btt_gatt_long_read(struct net_buf_simple *result_data, uint16_t *result_size,
-		       struct bt_conn *conn, enum bt_att_chan_opt bearer, uint16_t handle,
-		       uint16_t offset);
+int bt_testlib_gatt_long_read(struct net_buf_simple *result_data, uint16_t *result_size,
+			      struct bt_conn *conn, enum bt_att_chan_opt bearer, uint16_t handle,
+			      uint16_t offset);
 
 int bt_testlib_gatt_discover_primary(uint16_t *result_handle, uint16_t *result_end_handle,
 				     struct bt_conn *conn, const struct bt_uuid *uuid,
@@ -37,3 +40,5 @@ int bt_testlib_gatt_discover_characteristic(uint16_t *const result_value_handle,
 					    uint16_t *const result_def_handle, struct bt_conn *conn,
 					    const struct bt_uuid *uuid, uint16_t start_handle,
 					    uint16_t svc_end_handle);
+
+#endif /* ZEPHYR_TESTS_BLUETOOTH_COMMON_TESTLIB_INCLUDE_TESTLIB_ATT_READ_H_ */
