+++ tests/bluetooth/common/testlib/include/testlib/connect.h
@@ -1,0 +1,12 @@
+/* Copyright (c) 2023 Nordic Semiconductor ASA
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef ZEPHYR_TESTS_BLUETOOTH_COMMON_TESTLIB_INCLUDE_TESTLIB_CONNECT_H_
+#define ZEPHYR_TESTS_BLUETOOTH_COMMON_TESTLIB_INCLUDE_TESTLIB_CONNECT_H_
+
+#include <zephyr/bluetooth/conn.h>
+
+int bt_testlib_connect(const bt_addr_le_t *peer, struct bt_conn **conn);
+
+#endif /* ZEPHYR_TESTS_BLUETOOTH_COMMON_TESTLIB_INCLUDE_TESTLIB_CONNECT_H_ */
