+++ tests/bsim/bluetooth/host/att/long_read/CMakeLists.txt
@@ -5,20 +5,19 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr HINTS $ENV{ZEPHYR_BASE})
 project(app)
 
+add_subdirectory(${ZEPHYR_BASE}/tests/bluetooth/common/testlib testlib)
+
 target_sources(app PRIVATE
-  testlib/bs_main.c
-  testlib/adv.c
-  testlib/connect.c
-  testlib/scan.c
-  testlib/security.c
-  testlib/att_read.c
-  testlib/att_write.c
-  testlib/bs_sync.c
-  testlib/conn_wait.c
+  bs_main.c
+  bs_sync.c
   main.c
 )
 
 zephyr_include_directories(
   ${BSIM_COMPONENTS_PATH}/libPhyComv1/src/
   ${BSIM_COMPONENTS_PATH}/libUtilv1/src/
+)
+
+target_link_libraries(app PRIVATE
+  testlib
 )
