+++ tests/bsim/bluetooth/host/att/long_read/bs_sync.h
