--- tests/bsim/bluetooth/host/att/long_read/testlib/conn_wait.h
@@ -1,4 +1,0 @@
-#include <zephyr/bluetooth/conn.h>
-
-int bt_testlib_wait_connected(struct bt_conn *conn);
-int bt_testlib_wait_disconnected(struct bt_conn *conn);
