--- tests/bsim/bluetooth/host/att/long_read/testlib/scan.h
@@ -1,7 +1,0 @@
-/* Copyright (c) 2023 Nordic Semiconductor ASA
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#include <zephyr/bluetooth/bluetooth.h>
-
-int bt_testlib_scan_find_name(bt_addr_le_t *result, char name[]);
